Set user-agent on connector
The Connector needs the Agent name and version in the User-Agent header, but the Agent requires a valid Connector be provided in the constructor. The best current option is to have the Agent set the User-Agent string on the Connector during construction.

This breaks the tight coupling fo the Metadata class to the Agent class and enables sub-classes or alternate implementations of the ApmAgent to be properly reflected in the metadata.